+++ src/lib/pages/tx-details/components/balance-changes/balance-changes-nft.tsx
@@ -23,8 +23,14 @@ export const BalanceChangeNft = ({
   const nftObject = {
     collectionAddress: zAddr.optional().parse(nftMetadata?.collectionAddress),
     nftAddress: zHexAddr32.parse(formatAddresses(id).hex),
-    tokenId: nftMetadata?.tokenId,
-    uri: nftMetadata?.tokenUri,
+    tokenId:
+      nftMetadata?.tokenId && typeof nftMetadata.tokenId === "string"
+        ? nftMetadata.tokenId
+        : undefined,
+    uri:
+      nftMetadata?.tokenUri && typeof nftMetadata.tokenUri === "string"
+        ? nftMetadata.tokenUri
+        : undefined,
   };
   const { data: nft } = useNftMetadata(nftObject);
   const nftImage = useNftGlyphImage(nftObject);